move non-training specific modules out of training
