+++ sae_lens/__init__.py
@@ -3,20 +3,23 @@ __version__ = "2.1.3"
 
 from .analysis.hooked_sae_transformer import HookedSAETransformer
 from .cache_activations_runner import CacheActivationsRunner
-from .pretokenize_runner import pretokenize_runner
-from .sae_training_runner import SAETrainingRunner
-from .training.activations_store import ActivationsStore
-from .training.config import (
+from .config import (
     CacheActivationsRunnerConfig,
     LanguageModelSAERunnerConfig,
     PretokenizeRunnerConfig,
 )
-from .training.evals import run_evals
-from .training.sae import SAE, TrainingSAE
+from .evals import run_evals
+from .pretokenize_runner import pretokenize_runner
+from .sae import SAE, SAEConfig
+from .sae_training_runner import SAETrainingRunner
+from .training.activations_store import ActivationsStore
+from .training.training_sae import TrainingSAE, TrainingSAEConfig
 
 __all__ = [
     "SAE",
+    "SAEConfig",
     "TrainingSAE",
+    "TrainingSAEConfig",
     "HookedSAETransformer",
     "ActivationsStore",
     "LanguageModelSAERunnerConfig",
