+++ tests/benchmark/test_cache_activations_runner.py
@@ -4,7 +4,7 @@ import shutil
 import torch
 
 from sae_lens.cache_activations_runner import CacheActivationsRunner
-from sae_lens.training.config import CacheActivationsRunnerConfig
+from sae_lens.config import CacheActivationsRunnerConfig
 
 os.environ["WANDB_MODE"] = "offline"  # turn this off if you want to see the output
