+++ tests/benchmark/test_language_model_sae_runner.py
@@ -1,7 +1,7 @@
 import torch
 
+from sae_lens.config import LanguageModelSAERunnerConfig
 from sae_lens.sae_training_runner import SAETrainingRunner
-from sae_lens.training.config import LanguageModelSAERunnerConfig
 
 # os.environ["WANDB_MODE"] = "offline"  # turn this off if you want to see the output
