+++ tests/unit/training/test_activation_functions.py
@@ -1,7 +1,7 @@
 import pytest
 import torch
 
-from sae_lens.training.sae import get_activation_fn
+from sae_lens.sae import get_activation_fn
 
 
 def test_get_activation_fn_tanh_relu():
