+++ tests/unit/training/test_cache_activations_runner.py
@@ -9,11 +9,8 @@ from safetensors import safe_open
 from transformer_lens import HookedTransformer
 
 from sae_lens.cache_activations_runner import CacheActivationsRunner
+from sae_lens.config import CacheActivationsRunnerConfig, LanguageModelSAERunnerConfig
 from sae_lens.training.activations_store import ActivationsStore
-from sae_lens.training.config import (
-    CacheActivationsRunnerConfig,
-    LanguageModelSAERunnerConfig,
-)
 
 
 # The way to run this with this command:
