+++ tests/unit/training/test_config.py
@@ -1,7 +1,7 @@
 import pytest
 
 from sae_lens import __version__
-from sae_lens.training.config import LanguageModelSAERunnerConfig
+from sae_lens.config import LanguageModelSAERunnerConfig
 
 TINYSTORIES_MODEL = "tiny-stories-1M"
 TINYSTORIES_DATASET = "roneneldan/TinyStories"
