+++ tests/unit/training/test_evals.py
@@ -2,10 +2,11 @@ import pytest
 from datasets import Dataset
 from transformer_lens import HookedTransformer
 
+from sae_lens.config import LanguageModelSAERunnerConfig
+from sae_lens.evals import run_evals
+from sae_lens.sae import SAE
 from sae_lens.training.activations_store import ActivationsStore
-from sae_lens.training.config import LanguageModelSAERunnerConfig
-from sae_lens.training.evals import run_evals
-from sae_lens.training.sae import SAE, TrainingSAE
+from sae_lens.training.training_sae import TrainingSAE
 from tests.unit.helpers import TINYSTORIES_MODEL, build_sae_cfg, load_model_cached
