+++ tests/unit/training/test_sae_trainer.py
@@ -7,15 +7,15 @@ from datasets import Dataset
 from transformer_lens import HookedTransformer
 
 from sae_lens import __version__
+from sae_lens.config import LanguageModelSAERunnerConfig
 from sae_lens.training.activations_store import ActivationsStore
-from sae_lens.training.config import LanguageModelSAERunnerConfig
-from sae_lens.training.sae import TrainingSAE
 from sae_lens.training.sae_trainer import (
     SAETrainer,
     TrainStepOutput,
     _log_feature_sparsity,
     _update_sae_lens_training_version,
 )
+from sae_lens.training.training_sae import TrainingSAE
 from tests.unit.helpers import TINYSTORIES_MODEL, build_sae_cfg, load_model_cached
