+++ tests/unit/training/test_sparse_autoencoder_initialization.py
@@ -1,7 +1,7 @@
 import pytest
 import torch
 
-from sae_lens.training.sae import TrainingSAE
+from sae_lens.training.training_sae import TrainingSAE
 from tests.unit.helpers import build_sae_cfg
